requirements.txt
pandas==2.1.4
google-generativeai==0.3.2
pyyaml==6.0.1
pytest==7.4.3
flake8==6.1.0
